Student ID- 700744508
https://drive.google.com/file/d/1IqwJajRprn3hhmGkvoMAsTrVYsVttCRt/view?usp=share_link
Program 1- Creating class Employee with name, family, salary and department and adding a data member to count the number of Employees , defining average salary for the employee using a for loop. Creating a subclass of Employee i.e., Full time employee and defining it. Defining the main function by entering the employee details and printing the average salary of them.
Program 2-Installing a numpy function and importing the numpy as np. Creating a vector of size 20 with the values between 1 and 20. Reshaping the array to 4 by 5 and printing the output of first matrix. Later replacing the maximum number with zero and printing the second matrix.